Move transport tests into realtime suite
Also requires a change to Defaults.java to allow its tests to be in a
different package.
